Regroup webgl choropleth under one index export and start moving functions in utils
