+++ packages/visualizations/src/components/Map/WebGl/ChoroplethGeoJson.svelte
@@ -5,21 +5,20 @@
     import { debounce } from 'lodash';
     import type { SourceSpecification } from 'maplibre-gl';
     import type { BBox } from 'geojson';
-    import type { ColorScale, DataBounds, Color } from '../../../types';
-    import MapRender from '../MapRender.svelte';
-    import { BLANK } from '../../mapStyles';
-    import { getDataBounds, mapKeyToColor, VOID_BOUNDS } from '../../utils';
-    import { DEFAULT_COLORS, DEFAULT_COLORS_SCALE } from '../../constants';
+    import type { ColorScale, DataBounds, Color } from '../../types';
+    import MapRender from './MapRender.svelte';
+    import { BLANK } from '../mapStyles';
+    import { getDataBounds, mapKeyToColor, VOID_BOUNDS, computeFilterExpression, defaultFormat } from '../utils';
+    import { DEFAULT_COLORS, DEFAULT_COLORS_SCALE } from '../constants';
     import type {
         ChoroplethDataValue,
         ChoroplethLayer,
         GeoJsonChoroplethOptions,
         MapRenderTooltipFunction,
         ChoroplethShapeGeoJsonValue,
-        ChoroplethTooltipFormatter,
         MapLegend,
         MapFilter,
-    } from '../../types';
+    } from '../types';
 
     export let data: { value: ChoroplethDataValue[] }; // values, and the key to match
     export let options: GeoJsonChoroplethOptions; // contains the shapes to display & match
@@ -66,6 +65,7 @@
         newColorScale: ColorScale,
         values: ChoroplethDataValue[] = []
     ) {
+        // Specific to move outside ? On Call ?
         if (!newShapes.geoJson) {
             // We don't have everything we need yet
             return;
@@ -92,7 +92,7 @@
                 'fill-outline-color': DEFAULT_COLORS.ShapeOutline,
             },
         };
-
+        // Specific
         if (newShapes.geoJson) {
             source = {
                 type: 'geojson',
@@ -100,15 +100,12 @@
             };
 
             layer = baseLayer;
-
+            // Specific
             bbox = bbox || turfBbox(newShapes.geoJson) || VOID_BOUNDS;
         }
     }
 
     $: computeSourceLayerAndBboxes(shapes, colorScale, data.value);
-
-    const defaultFormat: ChoroplethTooltipFormatter = ({ value, label }) =>
-        value ? `${label} &mdash; ${value}` : label;
 
     $: renderTooltip = debounce(
         (hoveredFeature) => {
@@ -148,16 +145,6 @@
         10,
         { leading: true }
     );
-
-    function computeFilterExpression(filterConfig: MapFilter) {
-        /** Transform a filter object from the options into a Maplibre filter expression */
-        const { key, value } = filterConfig;
-        const filterMatchExpression: string[] = ['in', key];
-        (Array.isArray(value) ? value : [value]).forEach((filterValue) => {
-            filterMatchExpression.push(filterValue.toString());
-        });
-        return filterMatchExpression;
-    }
 
     $: if (filter) {
         filterExpression = computeFilterExpression(filter);
