--- packages/visualizations/src/components/Map/WebGl/VectorTiles/index.ts
@@ -1,12 +1,0 @@
-import type { DataFrame } from '../../../types';
-import type { ChoroplethOptions } from '../../types';
-import ChoroplethImpl from './Choropleth.svelte';
-import SvelteImpl from '../../../SvelteImpl';
-
-import 'maplibre-gl/dist/maplibre-gl.css';
-
-export default class VectorTilesChoropleth extends SvelteImpl<DataFrame, ChoroplethOptions> {
-    protected getSvelteComponentClass(): typeof ChoroplethImpl {
-        return ChoroplethImpl;
-    }
-}
