+++ packages/visualizations/src/index.ts
@@ -1,8 +1,7 @@
 export { default as Chart, _ChartJs, _ChartDataLabels } from './components/Chart';
 export { default as MarkdownText } from './components/MarkdownText';
 export { default as KpiCard } from './components/KpiCard';
-export { default as GeoJsonChoropleth } from './components/Map/WebGl/GeoJson';
-export { default as VectorTilesChoropleth } from './components/Map/WebGl/VectorTiles';
+export { GeoJsonChoropleth, VectorTilesChoropleth } from './components/Map/WebGl';
 export { default as SvgChoropleth } from './components/Map/Svg';
 export * from './types';
 export * from './components/types';
